move functionality that's not related to Edge.vertex into RoadView
